+++ src/main/php/response/format/HtmlFormatter.php
@@ -8,6 +8,7 @@
  * @package  stubbles\webapp
  */
 namespace stubbles\webapp\response\format;
+use stubbles\webapp\response\Headers;
 /**
  * Formats resource as HTML.
  *
@@ -97,10 +98,11 @@ class HtmlFormatter implements Formatter
     /**
      * formats resource for response
      *
-     * @param   mixed   $resource
+     * @param   mixed    $resource  resource data to create a representation of
+     * @param   Headers  $headers   list of headers for the response
      * @return  string
      */
-    public function format($resource)
+    public function format($resource, Headers $headers)
     {
         if (is_array($resource)) {
             if (isset($resource['title'])) {
@@ -165,7 +167,7 @@ class HtmlFormatter implements Formatter
     /**
      * write error message about 500 Internal Server error
      *
-     * @param   string  $message
+     * @param   string  $message  error messsage to display
      * @return  string
      */
     public function formatInternalServerError($message)