+++ src/main/php/response/format/VoidFormatter.php
@@ -8,6 +8,7 @@
  * @package  stubbles\webapp
  */
 namespace stubbles\webapp\response\format;
+use stubbles\webapp\response\Headers;
 /**
  * Formats resource to empty string.
  *
@@ -18,10 +19,11 @@ class VoidFormatter implements Formatter
     /**
      * formats resource for response
      *
-     * @param   mixed   $resource
+     * @param   mixed    $resource  resource data to create a representation of
+     * @param   Headers  $headers   list of headers for the response
      * @return  string
      */
-    public function format($resource)
+    public function format($resource, Headers $headers)
     {
         return '';
     }
@@ -61,7 +63,7 @@ class VoidFormatter implements Formatter
     /**
      * write error message about 500 Internal Server error
      *
-     * @param   string  $message
+     * @param   string  $message  error messsage to display
      * @return  string
      */
     public function formatInternalServerError($message)